+++ modules/apps/commerce/commerce-inventory-test/src/testIntegration/java/com/liferay/commerce/inventory/engine/test/CommerceInventoryEngineTest.java
@@ -366,8 +366,9 @@ public class CommerceInventoryEngineTest {
 		CommerceOrderItem commerceOrderItem =
 			_commerceOrderItemLocalService.addCommerceOrderItem(
 				_user.getUserId(), commerceOrder.getCommerceOrderId(),
-				_cpInstance1.getCPInstanceId(), null, bookQuantity, 0, 0,
-				StringPool.BLANK, _commerceContext, _serviceContext);
+				_cpInstance1.getCPInstanceId(), null,
+				BigDecimal.valueOf(bookQuantity), 0, 0, StringPool.BLANK,
+				_commerceContext, _serviceContext);
 
 		CommerceInventoryBookedQuantity commerceBookedQuantity =
 			_commerceBookedQuantityLocalService.addCommerceBookedQuantity(
