+++ modules/apps/commerce/commerce-inventory-test/src/testIntegration/java/com/liferay/commerce/inventory/search/test/CommerceInventoryBookedQuantityIndexerTest.java
@@ -53,6 +53,8 @@ import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
+import java.math.BigDecimal;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -172,8 +174,8 @@ public class CommerceInventoryBookedQuantityIndexerTest {
 		CommerceOrderItem commerceOrderItem =
 			_commerceOrderItemLocalService.addCommerceOrderItem(
 				_user.getUserId(), commerceOrder.getCommerceOrderId(),
-				cpInstance.getCPInstanceId(), null, 2, 0, 0, StringPool.BLANK,
-				_commerceContext, _serviceContext);
+				cpInstance.getCPInstanceId(), null, BigDecimal.valueOf(2), 0, 0,
+				StringPool.BLANK, _commerceContext, _serviceContext);
 
 		CommerceInventoryBookedQuantity commerceInventoryBookedQuantity =
 			_commerceBookedQuantityLocalService.addCommerceBookedQuantity(
@@ -228,8 +230,8 @@ public class CommerceInventoryBookedQuantityIndexerTest {
 		CommerceOrderItem commerceOrderItem =
 			_commerceOrderItemLocalService.addCommerceOrderItem(
 				_user.getUserId(), commerceOrder.getCommerceOrderId(),
-				cpInstance.getCPInstanceId(), null, 2, 0, 0, StringPool.BLANK,
-				_commerceContext, _serviceContext);
+				cpInstance.getCPInstanceId(), null, BigDecimal.valueOf(2), 0, 0,
+				StringPool.BLANK, _commerceContext, _serviceContext);
 
 		CommerceInventoryBookedQuantity commerceInventoryBookedQuantity =
 			_commerceBookedQuantityLocalService.addCommerceBookedQuantity(
