+++ modules/apps/commerce/commerce-media-test/src/testIntegration/java/com/liferay/commerce/media/test/CommerceMediaResolverTest.java
@@ -184,8 +184,8 @@ public class CommerceMediaResolverTest {
 		CommerceOrderItem commerceOrderItem =
 			_commerceOrderItemLocalService.addCommerceOrderItem(
 				_user.getUserId(), _commerceOrder.getCommerceOrderId(),
-				cpInstance.getCPInstanceId(), null, quantity, 0, quantity,
-				StringPool.BLANK,
+				cpInstance.getCPInstanceId(), null,
+				BigDecimal.valueOf(quantity), 0, quantity, StringPool.BLANK,
 				new TestCommerceContext(
 					_accountEntry, _commerceCurrency, _commerceChannel, _user,
 					_group, _commerceOrder),
